fixing range bug and cleaning unused code
